+++ packages/evm-address-tx-extractor/tests/extractor/mocked/ethers.mock.ts
@@ -1,0 +1,5 @@
+import { vi } from 'vitest';
+
+export const rpcInstance = {
+  getTransactionCount: vi.fn(),
+};