+++ packages/raw-data-provider/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @rosen-bridge/raw-data-provider
 
+## 0.1.7
+
+### Patch Changes
+
+- Update dependencies
+  - @rosen-bridge/abstract-observation-extractor@1.0.8
+  - @rosen-bridge/abstract-scanner@2.0.1
+  - @rosen-bridge/bitcoin-observation-extractor@8.0.3
+  - @rosen-bridge/bitcoin-runes-observation-extractor@3.0.4
+  - @rosen-bridge/bitcoin-scanner@1.1.1
+  - @rosen-bridge/cardano-observation-extractor@3.0.3
+  - @rosen-bridge/ergo-observation-extractor@2.0.3
+  - @rosen-bridge/evm-observation-extractor@7.0.3
+
 ## 0.1.6
 
 ### Patch Changes