+++ packages/reward-extractor/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @rosen-bridge/reward-extractor
 
+## 0.0.3
+
+### Patch Changes
+
+- Update dependencies
+  - @rosen-bridge/abstract-extractor@3.2.1
+
 ## 0.0.2
 
 ### Patch Changes